Put tests into their own source file.
