+++ Makefile
@@ -1,7 +1,7 @@
 all: sds-test
 
-sds-test: sds.c sds.h testhelp.h
-	$(CC) -o sds-test sds.c -Wall -std=c99 -pedantic -O2 -DSDS_TEST_MAIN
+sds-test: sds.c sds.h testhelp.h sds-test.c
+	$(CC) -o sds-test sds-test.c sds.c -Wall -std=c99 -pedantic -O2 -DSDS_TEST_MAIN
 	@echo ">>> Type ./sds-test to run the sds.c unit tests."
 
 clean:
